refactor: move tests to shared package
These tests are both related to the process instance expression context.
It does make sense that we package them together in a singular process
instance package. If we later extend the number of tests for the process
instance context, then we can just do so in this package. We don't have
to pollute the rest of the context tests.
