+++ zeebe/engine/src/test/java/io/camunda/zeebe/engine/processing/expression/processinstance/ProcessInstanceExpressionContextTest.java
@@ -5,7 +5,7 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.zeebe.engine.processing.expression;
+package io.camunda.zeebe.engine.processing.expression.processinstance;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.tuple;
